feat: 마이페이지 화면 구성 변경
