+++ app/src/main/java/net/developia/livartc/mypage/AccountFragment.kt
@@ -1,0 +1,7 @@
+package net.developia.livartc.mypage
+
+// TODO: Rename parameter arguments, choose names that match
+// the fragment initialization parameters, e.g. ARG_ITEM_NUMBER
+private const val ARG_PARAM1 = "param1"
+private const val ARG_PARAM2 = "param2"
+
