+++ app/src/main/java/net/developia/livartc/mypage/ViewPagerAdapter.kt
@@ -3,8 +3,8 @@ package net.developia.livartc.mypage
 import androidx.fragment.app.Fragment
 import androidx.fragment.app.FragmentActivity
 import androidx.viewpager2.adapter.FragmentStateAdapter
+import net.developia.livartc.mypage.account.AccountFragment
 import net.developia.livartc.mypage.badge.BadgeFragment
-import net.developia.livartc.mypage.coupon.CouponFragment
 import net.developia.livartc.mypage.membership.MemberShipFragment
 
 /**
@@ -19,7 +19,7 @@ class ViewPagerAdapter(fragment: FragmentActivity) : FragmentStateAdapter(fragme
         return when (position) {
             0 -> MemberShipFragment()
             1 -> BadgeFragment()
-            else -> CouponFragment()
+            else -> AccountFragment()
         }
     }
 }
