+++ app/src/main/java/net/developia/livartc/mypage/membership/MemberShipFragment.kt
@@ -45,37 +45,6 @@ class MemberShipFragment : Fragment() {
             val result = fetchDataFromServer()
             updateUI(result)
         }
-        // 1. my_info 버튼 클릭시 새로운 액티비티로 이동
-        val myInfoLayout = view.findViewById<LinearLayout>(R.id.my_info)
-        myInfoLayout.setOnClickListener {
-            startActivity(Intent(requireContext(), MyInfoActivity::class.java))
-        }
-
-        // 2. my_review 버튼 클릭시 새로운 액티비티로 이동
-        val myReviewLayout = view.findViewById<LinearLayout>(R.id.my_review)
-        myReviewLayout.setOnClickListener {
-            startActivity(Intent(requireContext(), MyReviewActivity::class.java))
-        }
-
-        // 3. my_review 버튼 클릭시 새로운 액티비티로 이동
-        val myOrderLayout = view.findViewById<LinearLayout>(R.id.my_order)
-        myOrderLayout.setOnClickListener {
-            startActivity(Intent(requireContext(), MyOrderActivity::class.java))
-        }
-
-        // 4. logout 버튼 클릭시 => 로그아웃하고 처음 화면으로 이동하기
-        // 일단 팝업 뜨는 버튼으로!!
-        val logoutLayout = view.findViewById<LinearLayout>(R.id.logout)
-        logoutLayout.setOnClickListener {
-            // startActivity(Intent(requireContext(), LogoutActivity::class.java))
-            // 팝업 뜨는 예시
-            val tmp = PopUpDto (
-                "Commenter",
-                "https://github.com/livarter/android/assets/77563814/a346484d-31df-477e-a43c-690aec02c63f",
-                "축하합니다! 뱃지를 발급하였습니다!"
-            )
-            PopUp().show(tmp, requireContext())
-        }
         return view
     }
