--- app/src/main/res/layout/fragment_coupon.xml
@@ -1,19 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-<ScrollView xmlns:android="http://schemas.android.com/apk/res/android"
-    xmlns:app="http://schemas.android.com/apk/res-auto"
-    xmlns:tools="http://schemas.android.com/tools"
-    android:layout_width="match_parent"
-    android:layout_height="match_parent"
-    android:background="@color/white"
-    tools:context=".mypage.coupon.CouponFragment" >
-
-
-    <LinearLayout
-        android:id="@+id/coupon_container"
-        android:layout_width="match_parent"
-        android:layout_height="wrap_content"
-        android:orientation="vertical">
-        <!-- 동적으로 추가될 뷰들 여기에 추가 -->
-    </LinearLayout>
-
-</ScrollView>
